+++ Serializers/WritableObjects/Writer/IWriter.cs
@@ -4,8 +4,8 @@ namespace Utils.Serializers.WritableObjects
 {
 	public interface IWriter : IDisposable
 	{
-		int Size { get; }
-		int Capacity { get; }
+		long Size { get; }
+		long Capacity { get; }
 		void Write<T>(T value);
 	}
 }
